this time i did it

+++ FVCalc.java
@@ -7,9 +7,9 @@ public class FVCalc {
 			return;
 		}
 
-		double money = Double.parseDouble(args[0]);
-		double rate = Double.parseDouble(args[1]);
-		double years = Double.parseDouble(args[2]);
+		int money = Integer.parseInt(args[0]);
+		int rate = Integer.parseInt(args[1]);
+		int years = Integer.parseInt(args[2]);
 
 		double finalValue = money * Math.pow((1 + rate/100), years);
 		System.out.println("after " + years + "years, " + money +
